refactor(join): extract shared HouseholdPreviewCard, kill the dialog drift
